Add qtum token
